requirements.txt
alpaca-py
python-dotenv
pandas
numpy
scikit-learn
stable-baselines3
streamlit
plotly
matplotlib
ta